WIP: clk: meson: g12a: fix hifi pll lock
The HIFI pll of the g12a sometimes takes a long time to report the lock in
HIFI_PLL_CNTL0 bit 31. The would eventually be reported but the delay may
be so long that the driver consider it a lock failure.

Bit 30 seems to do the same job but more quickly, let's try this instead.

Fixes: 085a4ea93d54 ("clk: meson: g12a: add peripheral clock controller")
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>